Adding exceptions, changing time inputs
BabySitter is adding exceptions onto the verifying methods
OutOfBoundsException was created because the preset exceptions didn't seem to align with what was happening
SetTimes was focused on changing the structure of the intake of time.  it is now a number from 5-11 PM or 12-4 AM.
BabySitterTest is updating test so that when setting the time it is the new format. Also throwing the exception.
SetTimesTest added test to create the transformTime method on SetTime.